코드 추가

+++ Projects/django/instagram/django_app/post/views/post.py
@@ -44,7 +44,8 @@ __all__ = (
 
 
 def post_list(request):
-    posts = Post.objects.all()
+    # posts = Post.objects.filter(is_visible=True)
+    posts = Post.visible.all()
     context = {
         'posts': posts,
     }
@@ -83,11 +84,16 @@ def post_add(request):
     return render(request, 'post/post_add.html', context)
 
 
-def post_delete(request, post_id):
+def post_delete(request, post_id, db_delete=False):
     if request.method == 'POST':
         post = Post.objects.get(id=post_id)
         if post.author.id == request.user.id:
-            post.delete()
+            if db_delete:
+                post.delete()
+            else:
+                post.is_visible = False
+                post.save()
+
         return redirect('post:list')
